+++ frontend/src/app/main/data/shortcuts.cljs
@@ -39,6 +39,16 @@
     (str "command+" shortcut)
     (str "ctrl+" shortcut)))
 
+(defn a-mod
+  "Adds the alt/option modifier to a shortcuts depending on the
+  operating system for the user"
+  [shortcut]
+  (str "alt+" shortcut))
+
+(defn ca-mod
+  [shortcut]
+  (c-mod (a-mod shortcut)))
+
 (defn bind-shortcuts [shortcuts bind-fn cb-fn]
   (doseq [[key {:keys [command disabled fn]}] shortcuts]
     (when-not disabled
@@ -61,8 +71,18 @@
      "Shift+")
    key))
 
+(defn alt [key]
+  (str
+   (if (cfg/check-platform? :macos)
+     mac-option
+     "Alt+")
+   key))
+
 (defn meta-shift [key]
   (-> key meta shift))
+
+(defn meta-alt [key]
+  (-> key meta alt))
 
 (defn supr []
   (if (cfg/check-platform? :macos)